MOD: perfil items

+++ app/src/main/java/com/bunizz/instapetts/db/Utilities.java
@@ -165,10 +165,10 @@ public class Utilities {
     }
 
 
-    public static float convertDpToPixel(float dp, Context context) {
+    public static int convertDpToPixel(float dp, Context context) {
         Resources resources = context.getResources();
         DisplayMetrics metrics = resources.getDisplayMetrics();
-        float px = dp * ((float) metrics.densityDpi / DisplayMetrics.DENSITY_DEFAULT);
+        int px = (int) (dp * ( metrics.densityDpi / DisplayMetrics.DENSITY_DEFAULT));
         return px;
     }
